+++ storage/s3/upload.go
@@ -12,7 +12,6 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/s3"
 
 	"gitlab.com/nunet/device-management-service/models"
-	"gitlab.com/nunet/device-management-service/storage"
 	"gitlab.com/nunet/device-management-service/storage/basic_controller"
 )
 
@@ -22,7 +21,7 @@ import (
 // Warning: the implementation should rely on the FS provided by the volume controller,
 // be careful if managing files with `os` (the volume controller might be
 // using an in-memory one)
-func (s *S3Storage) Upload(ctx context.Context, vol storage.StorageVolume,
+func (s *S3Storage) Upload(ctx context.Context, vol models.StorageVolume,
 	destinationSpecs *models.SpecConfig) error {
 
 	target, err := DecodeInputSpec(destinationSpecs)